chore(readme): Update contributing guidelines link
*Description of changes:*
NPM Artifacts do not contain CONTRIBUTING.md and so this relative link will not resolve. Changing it to an absolute URL means we do not need to increase the artifact size. 

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.